moved ImageViewerDialog
