+++ src/containers/index.ts
@@ -1,11 +1,9 @@
 import ConnectedApplication from './ConnectedApplication';
 import ConnectedDeleteImageDialog from './ConnectedDeleteImageDialog';
-import ConnectedImageViewer from './ConnectedImageViewer';
 import ConnectedImportImagesButton from './ConnectedImportImagesButton';
 
 export {
   ConnectedApplication,
   ConnectedDeleteImageDialog,
-  ConnectedImageViewer,
   ConnectedImportImagesButton
 };
