+++ src/pages/images/GalleryDialog/ImageViewerDialog/BrightnessSlider/BrightnessSlider.css.tsx
