+++ src/pages/images/GalleryDialog/ImageViewerDialog/BrightnessSlider/BrightnessSlider.tsx
