+++ src/pages/images/GalleryDialog/ImageViewerDialog/ContrastSlider/ContrastSlider.css.tsx
