+++ src/pages/images/GalleryDialog/ImageViewerDialog/ContrastSlider/ContrastSlider.tsx
