+++ src/pages/images/GalleryDialog/ImageViewerDialog/ImageHistogram/ImageHistogram.css.tsx
