+++ src/pages/images/GalleryDialog/ImageViewerDialog/ImageHistogram/ImageHistogram.js
