+++ src/pages/images/GalleryDialog/ImageViewerDialog/ImageViewer/ImageViewer.css.tsx
