+++ src/pages/images/GalleryDialog/ImageViewerDialog/ImageViewer/ImageViewer.tsx
@@ -9,8 +9,8 @@ import Tooltip from '@material-ui/core/Tooltip';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 import EqualizerIcon from '@material-ui/icons/Equalizer';
 import PublicIcon from '@material-ui/icons/Public';
-import { ImageViewerExposureDrawer } from '..';
-import Image from '../../../components/Image/Image';
+import { ImageViewerExposureDrawer } from '../index';
+import Image from '../../../../../components/Image/Image';
 import { makeStyles } from '@material-ui/styles';
 
 const useStyles = makeStyles(styles);
