+++ src/pages/images/GalleryDialog/ImageViewerDialog/ImageViewerDialog/ImageViewerDialog.css.tsx
